refactor(auth): split NextAuth config helpers
